Update running_tracker_classes_v4.py
In this version, I have fixed a few bugs and enabled the functionality for data strings, written in the GUI, to be uploaded in a ".csv" file. I still need to set a few parameters so that when the submit button is pressed, it opens the main interface. Additionally, in this script the new & existing user window is being skipped. Also, for more context, the script that creates the agenda for when the user should run needs to occur right after when the submit button is pressed.